silence warnings in unit tests

+++ unit/test_debruijn.c
@@ -5,7 +5,7 @@ static const char* debruijn_charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmno
 
 bool test_r_debruijn_pattern(void) {
 	char* pattern = r_debruijn_pattern (256, 0, debruijn_charset);
-	mu_assert_eq (strlen(pattern), 256, "pattern length");
+	mu_assert_eq ((int)strlen(pattern), 256, "pattern length");
 	mu_assert_streq (pattern, "AAABAACAADAAEAAFAAGAAHAAIAAJAAKAALAAMAANAAOAAPAAQAARAASAATAAUAAVAAWAAXAAYAAZAAaAAbAAcAAdAAeAAfAAgAAhAAiAAjAAkAAlAAmAAnAAoAApAAqAArAAsAAtAAuAAvAAwAAxAAyAAzAA1AA2AA3AA4AA5AA6AA7AA8AA9AA0ABBABCABDABEABFABGABHABIABJABKABLABMABNABOABPABQABRABSABTABUABVABWABXABY", "pattern of 256 length");
 	free (pattern);
 	mu_end;
